implemented all constructors for sparse, extended tests, need to fix << for unordered_map and Polynomial<T>, maybe switch to extrovert <<

+++ utils.hpp
@@ -6,7 +6,7 @@
 template <typename T>
 bool is_zero(T a)
 {
-    return (a == 0);
+    return (static_cast<int>(a) == 0);
 }
 
 template <typename T>
